+++ MachineState.h
@@ -10,15 +10,15 @@
 
 class MachineState {
     std::map<int, vips::VImage> slots;
-    std::map<int, int> variables;
+    std::map<int, double> variables;
 public:
     MachineState();
 
     void set_image(int index, vips::VImage image);
     vips::VImage get_image(int index);
     void free_image(int index);
-    void set_variable(int index, int value);
-    int get_variable(int index);
+    void set_variable(int index, double value);
+    double get_variable(int index);
 };
